+++ tests/migrations/m180322_213900_create_table_test_pk_composite.php
@@ -8,6 +8,9 @@ class m180322_213900_create_table_test_pk_composite extends Migration
 {
     public function up(): void
     {
+        if ($this->db->driverName === 'sqlite') {
+            return;
+        }
         $tableOptions = null;
         if ($this->db->driverName === 'mysql') {
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
@@ -23,6 +26,9 @@ class m180322_213900_create_table_test_pk_composite extends Migration
 
     public function down(): void
     {
+        if ($this->db->driverName === 'sqlite') {
+            return;
+        }
         $this->dropTable('{{%test_pk_composite}}');
     }
 }
