+++ tests/migrations/m180324_105400_create_table_test_fk.php
@@ -8,6 +8,9 @@ class m180324_105400_create_table_test_fk extends Migration
 {
     public function up(): void
     {
+        if ($this->db->driverName === 'sqlite') {
+            return;
+        }
         $tableOptions = null;
         if ($this->db->driverName === 'mysql') {
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
@@ -22,6 +25,9 @@ class m180324_105400_create_table_test_fk extends Migration
 
     public function down(): void
     {
+        if ($this->db->driverName === 'sqlite') {
+            return;
+        }
         $this->dropTable('{{%test_fk}}');
     }
 }
